(Refactoring) Move database instance to app file
