+++ Week9/.idea/libraries/Gradle__org_jetbrains_kotlinx_kotlinx_coroutines_test_1_3_8_jar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.8@jar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlinx/kotlinx-coroutines-test/1.3.8/14097b608fa5a7ea3de3dfef061aa715aa1cb7fe/kotlinx-coroutines-test-1.3.8.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlinx/kotlinx-coroutines-test/1.3.8/5064c6037402c6a106a8e54eb39067e12b152a60/kotlinx-coroutines-test-1.3.8-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlinx/kotlinx-coroutines-test/1.3.8/ab22bd4b4b34e6d19ac94ab0a0e7d742b3397a87/kotlinx-coroutines-test-1.3.8-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
